% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/html.table.reader.R
\name{html.table.reader}
\alias{html.table.reader}
\title{Read an Html Table from a url in an html.table file.}
\usage{
html.table.reader(data.file, filename, variable.name)
}
\arguments{
\item{data.file}{The name of the data file to be read.}

\item{filename}{The path to the data set to be loaded.}

\item{variable.name}{The name to be assigned to in the global environment.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function will head an HTMl file from the url,
parse the file for tables, and extract all tables
into dataframes that are names using the table id
}
\examples{
library('ProjectTemplate')

\dontrun{es.reader('example.es', 'data/example.es', 'example')}
}

